portal/requirements.txt
Flask==3.0.3
prometheus_client==0.21.0
pydantic[email]
Flask-Cors
Flask-RESTful
pytest==7.4.0
pytest-cov==4.1.0
black==23.9.0
isort==5.12.0
mypy==1.5.0
requests==2.31.0
python-dotenv==1.0.0
loguru==0.7.0
kubernetes 
